신호등 + 종료
