+++ 희승/08.14(신호등 + 종료)/left.png
